refactor: Move string newtypes -> mod str
Create a new module for strings with invariants.
